requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-writable-nested==0.7.0
inflection==0.5.1
jsonschema==4.17.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
uritemplate==4.1.1
